Refactor: Move all system-specific models to own module system_settings
